backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
prisma==0.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
httpx==0.26.0
python-multipart==0.0.6
PyMuPDF==1.23.21
Pillow==10.2.0
pytesseract==0.3.10
